Major python updates (shrinks, minipy py3, v1.1.0).
Update minipy to run with python3 (not just hacks to allow it to
compress python3 code). This is necessary for some of the newer syntax
supported by python3 such as {**foo, **bar} for merging maps.

Major updates to the python implementation to use the new macro
mechanism (no macroexpand dedicated function) and a bunch of other
refactoring to make the python implementation more closely match the JS
one.

The resulting improvement in size for stepA_miniMAL.pyz is from 1403
bytes to 1081 bytes (1059 without the shebang).